Version 1.0
Ability to post anonymously (no login required), view all posts,
register as a user, log in, and post as user. Basic private profile that
shows the users credentials.

The application also registers continuously the clients device from
which he is operating. In the case of a post, the device is inserted
into the database and is shown to all via the datatable, in relation the
the post that was done.