Small update to ACL pages.
I have introduced the other pages that display current users, roles and
permissions in the db, the ability to modify these entities is still
pending.

Modified the login to now request from the db the name of the users role
and store it in session data, this will reduce calls to the db later on.

I have also moved the read_db() function from the Permission_database
model to the Login_database model, due to its flexibility we can use
this to replace functions once stored in Login_database. This has made
the Permission_database model redundant and pending deletion.